[usb-moded-qt5] Moved constants from QUsbModed to base class QUsbMode. JB#26403
QUsbMode can be used for the purposes of exporting USB mode constants to QML.
Unlike QUsbModed, its instantiation doesn't involve any D-Bus action.
