+++ src/qusbmoded.h
@@ -34,14 +34,13 @@
 #ifndef QUSBMODED_H
 #define QUSBMODED_H
 
-#include "qusbmoded_types.h"
+#include "qusbmode.h"
 
-#include <QObject>
 #include <QStringList>
 
 class QDBusPendingCallWatcher;
 
-class QUSBMODED_EXPORT QUsbModed : public QObject
+class QUSBMODED_EXPORT QUsbModed : public QUsbMode
 {
     Q_OBJECT
     Q_PROPERTY(bool available READ available NOTIFY availableChanged)
@@ -49,53 +48,7 @@ class QUSBMODED_EXPORT QUsbModed : public QObject
     Q_PROPERTY(QString currentMode READ currentMode WRITE setCurrentMode NOTIFY currentModeChanged)
     Q_PROPERTY(QString configMode READ configMode WRITE setConfigMode NOTIFY configModeChanged)
 
-    // States (from usb_moded-dbus.h)
-    Q_PROPERTY(QString USB_CONNECTED READ USB_CONNECTED CONSTANT)
-    Q_PROPERTY(QString DATA_IN_USE READ DATA_IN_USE CONSTANT)
-    Q_PROPERTY(QString USB_DISCONNECTED READ USB_DISCONNECTED CONSTANT)
-    Q_PROPERTY(QString USB_CONNECTED_DIALOG_SHOW READ USB_CONNECTED_DIALOG_SHOW CONSTANT)
-
-    // Modes (from usb_moded-modes.h)
-    Q_PROPERTY(QString MODE_UNDEFINED READ MODE_UNDEFINED CONSTANT)
-    Q_PROPERTY(QString MODE_ASK READ MODE_ASK CONSTANT)
-    Q_PROPERTY(QString MODE_MASS_STORAGE READ MODE_MASS_STORAGE CONSTANT)
-    Q_PROPERTY(QString MODE_DEVELOPER READ MODE_DEVELOPER CONSTANT)
-    Q_PROPERTY(QString MODE_MTP READ MODE_MTP CONSTANT)
-    Q_PROPERTY(QString MODE_HOST READ MODE_HOST CONSTANT)
-    Q_PROPERTY(QString MODE_CONNECTION_SHARING READ MODE_CONNECTION_SHARING CONSTANT)
-    Q_PROPERTY(QString MODE_DIAG READ MODE_DIAG CONSTANT)
-    Q_PROPERTY(QString MODE_ADB READ MODE_ADB CONSTANT)
-    Q_PROPERTY(QString MODE_PC_SUITE READ MODE_PC_SUITE CONSTANT)
-    Q_PROPERTY(QString MODE_CHARGING READ MODE_CHARGING CONSTANT)
-    Q_PROPERTY(QString MODE_CHARGER READ MODE_CHARGER CONSTANT)
-
 public:
-    class Mode {
-    public:
-        // States (from usb_moded-dbus.h)
-        static const QString Connected;
-        static const QString DataInUse;
-        static const QString Disconnected;
-        static const QString ModeRequest;
-
-        // Modes (from usb_moded-modes.h)
-        static const QString Undefined;
-        static const QString Ask;
-        static const QString MassStorage;
-        static const QString Developer;
-        static const QString MTP;
-        static const QString Host;
-        static const QString ConnectionSharing;
-        static const QString Diag;
-        static const QString Adb;
-        static const QString PCSuite;
-        static const QString Charging;
-        static const QString Charger;
-
-    private:
-        Mode(); // Disallow instantiation
-    };
-
     explicit QUsbModed(QObject* parent = NULL);
     ~QUsbModed();
 
@@ -106,25 +59,6 @@ public:
 
     bool setCurrentMode(QString mode);
     bool setConfigMode(QString mode);
-
-private:
-    // Getters for QML constants
-    QString USB_CONNECTED() const { return Mode::Connected; }
-    QString DATA_IN_USE() const { return Mode::DataInUse; }
-    QString USB_DISCONNECTED() const { return Mode::Disconnected; }
-    QString USB_CONNECTED_DIALOG_SHOW() const { return Mode::ModeRequest; }
-    QString MODE_UNDEFINED() const { return Mode::Undefined; }
-    QString MODE_ASK() const { return Mode::Ask; }
-    QString MODE_MASS_STORAGE() const { return Mode::MassStorage; }
-    QString MODE_DEVELOPER() const { return Mode::Developer; }
-    QString MODE_MTP() const { return Mode::MTP; }
-    QString MODE_HOST() const { return Mode::Host; }
-    QString MODE_CONNECTION_SHARING() const { return Mode::ConnectionSharing; }
-    QString MODE_DIAG() const { return Mode::Diag; }
-    QString MODE_ADB() const { return Mode::Adb; }
-    QString MODE_PC_SUITE() const { return Mode::PCSuite; }
-    QString MODE_CHARGING() const { return Mode::Charging; }
-    QString MODE_CHARGER() const { return Mode::Charger; }
 
 Q_SIGNALS:
     void availableChanged();
